Update and rename isPalindrome.java to IsPalindrome.java
